Remove core.py
